Trees

+++ Source/Game/Models/Position.cs
@@ -42,7 +42,7 @@
             Y = yPosition;
         }
 
-        public bool InRange(int xLb, int yLb, int xUb, int yUb)
+        public bool InRange(float xLb, float yLb, float xUb, float yUb)
         {
             return (xLb <= this.X && this.X <= xUb) && (yLb <= this.Y && this.Y <= yUb);
         }
